third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-015-expected.txt


FAIL CSS Test: @viewport constrained - zoom less than min-zoom assert_equals: expected 2 but got 0.326531
